+++ src/components/admin/adminNotice/AdminNoticeList.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { SpinnerWrapperStyled } from '../../../../components/user/mypage/Spinner.styled';
+import { SpinnerWrapperStyled } from '../../user/mypage/Spinner.styled';
 
 export const SpinnerWrapper = styled(SpinnerWrapperStyled)`
   width: 100%;
