+++ src/components/common/admin/searchBar/SearchBar.styled.ts
