+++ src/pages/admin/adminNotice/AdminNotice.tsx
@@ -1,12 +1,5 @@
-import { Outlet } from 'react-router-dom';
-import AdminTitle from '../../../components/common/admin/title/AdminTitle';
-import * as S from './AdminNotice.styled';
+import CommonAdminPage from '../commonAdminPage';
 
 export default function AdminNotice() {
-  return (
-    <S.AdminNoticeContainer>
-      <AdminTitle title='공지사항' />
-      <Outlet />
-    </S.AdminNoticeContainer>
-  );
+  return <CommonAdminPage title='공지사항' />;
 }
