+++ src/pages/admin/adminNotice/adminNoticeList/AdminNoticeListPage.tsx
@@ -1,0 +1,5 @@
+import AdminNoticeList from '../../../../components/admin/adminNotice/AdminNoticeList';
+
+export default function AdminNoticeListPage() {
+  return <AdminNoticeList />;
+}
