+++ src/pages/admin/adminNotice/adminNoticeWrite/AdminNoticeWritePage.tsx
@@ -1,0 +1,5 @@
+import AdminNoticeWrite from '../../../../components/admin/adminNotice/AdminNoticeWrite';
+
+export default function AdminNoticeWritePage() {
+  return <AdminNoticeWrite />;
+}
